pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vectordb-migration"
version = "0.1.0"
description = "A library for migrating vector embeddings between different vector database systems"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["vector", "database", "migration", "pgvector", "qdrant", "embeddings"]
authors = [
    {name = "Original Author", email = "example@example.com"}
]
maintainers = [
    {name = "Current Maintainer", email = "example@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
]
dependencies = [
    "psycopg2-binary>=2.9.0",
    "qdrant-client>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/vectordb-migration"
"Bug Tracker" = "https://github.com/yourusername/vectordb-migration/issues"
"Documentation" = "https://vectordb-migration.readthedocs.io/"
"Source Code" = "https://github.com/yourusername/vectordb-migration"

[project.scripts]
vectordb-migrate = "vectordb_migration.cli:main"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["vectordb_migration"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
